refactor: 파일,변수명 변경
