--- src/modules/Login/index.ts
@@ -1,1 +1,0 @@
-export { default as loginReducer } from './reducer';
